chore: improve and move Directory class
