+++ .github/internal-cicd/update-workflow-versions.ts
@@ -2,7 +2,7 @@ import { TagClient, RepoClient } from "../../deps.ts";
 import { Input} from "../../deps.ts";
 import chalk from "../../deps.ts";
 import { File } from "../../cicd/core/File.ts";
-import { Directory } from "../../cicd/core/Directory.ts";
+import { Directory } from "../../deps.ts";
 import { Utils } from "../../cicd/core/Utils.ts";
 
 if (Deno.args.length != 2) {
