+++ deps.ts
@@ -4,6 +4,7 @@ import {
 } from "https://deno.land/x/kd_clients@v1.0.0-preview.4/GitHubClients/mod.ts";
 import { Input } from "https://deno.land/x/cliffy@v1.0.0-rc.3/prompt/input.ts";
 import chalk from "npm:chalk@5.3.0";
+import { Directory } from "./.github/internal-cicd/core/Directory.ts";
 
 export {
 	IssueClient, PullRequestClient, ProjectClient,
@@ -11,3 +12,4 @@ export {
 };
 export { Input };
 export default chalk;
+export { Directory };
